// src/app/shared/model/met-view.ts
export class MetView {
  supGuid?: string;
  supCustomerId?: string;
  supDataCategory?: string;
  supInvariantGuid?: string;
  supCreatedTimestamp?: string;
  supLastUpdateTimestamp?: string;
  supUniversalIndex?: string;
  supVersionNo?: string;
  supVersionEffectiveFrom?: string;
  supVersionEffectiveTo?: string;
  supId?: string;
  supShortTitle?: string;
  supTitle?: string;
  supDescription?: string;
  supAlias1ShortTitle?: string;
  supAlias2ShortTitle?: string;
  supAlias3ShortTitle?: string;
  supAlias1Title?: string;
  supAlias2Title?: string;
  supAlias3Title?: string;
  supAlias1Description?: string;
  supAlias2Description?: string;
  supAlias3Description?: string;
  supManagementName?: string;
  supInternalName?: string;
  supImplmentationName?: string;
  supUri?: string;
  id?: number;
  mcoDisplaySequence?: number;
  mcoIsActive?: true;
  mcoIsFactorySet?: false;
  mcoIsConstant?: false;
  mcoMlaId?: string;
  label?: null;
  tovMultiViewFormat?: string;
  tovUniViewFormat?: any;
  tovCentricity?: string;
  topology: any;
  orientation: any;
  includeSubject?: boolean;
  contentType: any;
  defaultSequenceDirection?: string;
  defaultSequenceSubject?: string;
  viewType?: string;
  temp: any;
  refCatViewGuid: any;
  refCatViewTitle: any;
  refCatViewSupid: any;
  datView: any;
  metMediaFile: any;
  metTypeOfProperty?: [];
}