// src/app/shared/model/category.ts
export class Category {
  supAlias1Description?: any;
  supAlias1ShortTitle?: any;
  supAlias1Title?: any;
  supAlias2Description?: any;
  supAlias2ShortTitle?: any;
  supAlias2Title?: any;
  supAlias3Description?: any;
  supAlias3ShortTitle?: any;
  supAlias3Title?: any;
  supCreatedTimestamp?: any;
  supCustomerId?: any;
  supDataCategory?: any;
  supDescription?: any;
  supGuid?: string;
  supId?: string;
  supImplmentationName?: any;
  supInternalName?: any;
  supInvariantGuid?: any;
  supLastUpdateTimestamp?: any;
  supManagementName?: any;
  supShortTitle?: any;
  supTitle?: string;
  supUniversalIndex?: any;
  supUri?: any;
  supVersionEffectiveFrom?: any;
  supVersionEffectiveTo?: any;
  supVersionNo?: any;
}